ai-service/requirements.txt

fastapi
uvicorn
anthropic
python-dotenv
